Move organization dux to opensrp directory
