+++ src/store/ducks/tests/fixtures.ts
@@ -1,4 +1,5 @@
 import { clone, cloneDeep } from 'lodash';
+import { Organization } from '../opensrp/organizations';
 import { InterventionType, Plan, PlanStatus } from '../plans';
 
 export const plan1: Plan = {
@@ -1178,7 +1179,7 @@ export const CompleteRoutineTableProps = {
   useDrillDownTrProps: false,
 };
 
-export const organization1: any = {
+export const organization1: Organization = {
   active: true,
   id: 1,
   identifier: 'fcc19470-d599-11e9-bb65-2a2ae2dbcce4',
@@ -1194,14 +1195,14 @@ export const organization1: any = {
   },
 };
 
-export const organization2: any = {
+export const organization2: Organization = {
   active: true,
   id: 3,
   identifier: '4c506c98-d3a9-11e9-bb65-2a2ae2dbcce4',
   name: 'Demo Team',
 };
 
-export const organization3: any = {
+export const organization3: Organization = {
   active: true,
   id: 2,
   identifier: 'd23f7350-d406-11e9-bb65-2a2ae2dbcce4',
@@ -1218,7 +1219,7 @@ export const organization3: any = {
   },
 };
 
-export const organizations: any[] = [organization1, organization2];
+export const organizations: Organization[] = [organization1, organization2];
 
 export const practitioner1 = {
   active: true,
